Modify headers and restructure project
