+++ examples/position_listener.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <housou/listener.hpp>
+#include <housou/housou.hpp>
 
 #include <stdlib.h>
 #include <unistd.h>
