+++ src/base_broadcaster.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <housou/base_broadcaster.hpp>
+#include <housou/broadcaster/base_broadcaster.hpp>
 
 #include <arpa/inet.h>
 #include <ifaddrs.h>
